backend/requirements.txt
quart
sqlalchemy
bcrypt
pyjwt

# core 模块依赖
onnxruntime-gpu  # GPU推理（无CUDA自动回退CPU）；纯CPU环境改为 onnxruntime
opencv-python
numpy
boxmot

# 测试依赖
pytest
pytest-asyncio
pytest-cov